+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Implemented enhancements
 - TT-1733 - Add request for individual verification feature
-
+- TT-1723 - Add collaborator list admin view
+  
 ### Fixed bugs:
 - TT-1728 - Not ask personal details to individual upgrading to business profile
 - No ticket - Remove obsolete settings